Pozabezpieczanie endpointów powiadomień
-> Dodanie weryfikacji tokenu

+++ Backend/MoviesWebApplication/Movies.Application/Notifications/DeleteNofitication.cs
@@ -1,12 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using MediatR;
+﻿using MediatR;
+using Microsoft.AspNetCore.Http;
+using Microsoft.EntityFrameworkCore;
 using Movies.Domain.Entities;
 using Movies.Infrastructure;
-using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Claims;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace Movies.Application.Notifications
 {
@@ -22,10 +24,12 @@ namespace Movies.Application.Notifications
     public class DeleteNotificationCommandHandler : IRequestHandler<DeleteNotification, Notification>
     {
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public DeleteNotificationCommandHandler(DataContext context)
+        public DeleteNotificationCommandHandler(DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         public async Task<Notification> Handle(DeleteNotification request, CancellationToken cancellationToken)
@@ -36,6 +40,13 @@ namespace Movies.Application.Notifications
             if (notification == null)
             {
                 return null;
+            }
+
+            var currentUserId = _httpContextAccessor.HttpContext?.User?.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (notification.TargetUserId != currentUserId)
+            {
+                throw new UnauthorizedAccessException("Nie masz uprawnień do usunięcia tego powiadomienia!");
             }
 
             _context.Notifications.Remove(notification);
